Render and Preview fix:
Rendering and preview of materials was not working with the latest
Blender svn rev 35000 and later.  This was due to Ribmosaic trying to modify
properties during rendering which is not allowed anymore.  The temporary fix is
having a dummy pass that the exporter can modify and use if the user has not
provided a render pass.

Also put in debug prints that output method names and important vars to the
console to help with debugging.  Yes, there are better ways to do this but
print statements are simple to
put in and modify easily.