Pipeline XML parsing fixes:
In Python 2.7 and beyond, ElementTree.GetChildren method is deprecated.
Now using iterfind('*') to get child elements.

If xml path is an empty string then assume that the root element tree is to
be used.

Too many "/" were in the xml path when parsing a category element.

Added debug print statements to help with debugging.