PipelineManager._unload_xml() bug fix:
calling _unlad_xml() would cause several python exceptions.
The method assumed a valid parent of an element would always be found.
If the element was the root element in a pipeline then parent would be
None.  The code now checks if parent is valid before using it.  If the parent
is not found then the pipeline tree root element is used as the parent.