Ribmosaic panel displays - continuous error messages fix:
After switching the render engine to "Rib Mosaic" there would be
a continuous stream of error messages in the console.  This was being
caused by a call to PipelineManager.sync() method during the panel's
poll execution.  Blender does not allow the modification of DataBlocks or
user defined properties during rendering or user interface calls
ie poll and draw.

  So for now the sync call gets executed when the addon gets registered and
can also be triggered by the user through the pipeline panel.