Bug fix for register attribute conflict:
Blender now allows operators and property classes to have a register()
and unregister() method.  When an addon gets enabled the classes register()
method will get called if it exists.

The problem was the WM_OT_ribmosaic_library_addpanel class had a property
called register (flag for panel registration state).  Blender doesn't check
to see if the class attribute "register" that it found is a method and blindly
tries to execute it and of course fails.

The fix was just to rename the property to panelregister.