Very basic object exporting working:
Objects in the Blender scene are now exported to the RIB but in a very
simple form.  For now the RIB data is put inline in the pass RIB.
No object data is exported.

Added some RIB formatting Helper functions which are take from
Matt Ebb's blender to 3delight exporter.