Framebuffer display disabled by default in aqsis core pipeline:
The utility panel for frame buffer display has been set to disabled as
the default setting.  The user can easily enable it from the pipeline manager
panel if they want to see the frame buffer display.

Disabled it since the preview would always use the frame buffer display
which is annoying and is a bug that needs to be fixed.