Bug Fix: creating new pipeline raises exception:
If the user tried to create a new pipeline in the Pipeline manager
panel then an exception was raised. _new_element() method was not
checking if the xmlpath was blank indicating that the root element
is to be used.

list_elements() method now correctly gets the root element if the
xmlpath is blank.