Update for Blender API change:
In Blender svn rev:36102 bpy.data.is_dirty now indicates if recent edits
have been saved to disk.  This meant the user now had to save every time
before doing a render or a preview which was a pain.

The new property bpy.data.is_saved indicates if the current session has
been saved to disk as a .blend file.  In other words a .blend file does
exist for the current session.  This is what the export manager wants
to check so that it can use the .blend path for export.