Bug Fix: ExportArchive.list_targets() was not doing proper comparison:
list_targets() was only using the file extension for the comparison instead
of checking if the match string was in the end of the file name.  If the
match string was more than just the file extension then no matches would
be found.

Now use the .endwith() string method to see if the match string is in the
end of the file name.