Bug Fix: too many vertices being exported:
Some .blend files have more vertices than what are used by faces.  This can
happen if vertices are not assigned to a face or when deleting faces but not
vertices.  This is evident in the .blend used for blender preview scene
rendering.

Having more vertices than what the faces are using genderates error messages
in Aqsis, Pixie, and 3Delight.

The fix was to find out the highest index of the vertice used and don't
export anything higher.

This fixes #3 issue on github repo.