ExporterArchive.export_shaders() method added:
There are several panels that need to export shaders of different types
so there is now a generic method for exporting shaders that are pinned
to a specific data block.  The method eliminates a lot of code duplication.

ExporterArchive.riTransform() method added to help with exporting a matrix transform.

A light now exports its transform automatically.