potential Bug Fix:
Don't use OS dependant path seperator in paths used in RIB.
Use only '/' for path seperator.  On linux this is not a problem since
the path seperator is '/'. On Windows the seperator is '\' which
ends up acting as an escape character when put in a string var in Aqsis.

get_archive_paths() and make_export_path() were modified to take a path
seperator parameter to allow for more flexible path seperator usage.

This bug fix needs to be verified by someone using Windows.
This change is in response to #13 issue on github repo:
https://github.com/nfz/RIBMosaic-exp/issues/13