Bug Fix: slmeta file name includes compiled name extension
the slmeta files should only have extension .slmeta and not .slx.slmeta
if compiled by aqsl.  This is so that generic pipline shader libraries can be
built and not be tied to a specific pipeline core ie Aqsis.

The Fix: Was just a matter of modifying the pipeline xml that uses a link
expression that replaces the compiler specific file extension with .slmeta.

This fixes #20 issue on GitHub: https://github.com/nfz/RIBMosaic-exp/issues/20