New Feature: shader in text editor can be added to active pipeline
A new blender operator WM_OT_ribmosaic_text_addshaderpanel() has been
added that allows the user to add shaders that are in the text editor to
the active pipeline by right mouse button click on the shader text and
selecting "Add Shader Panels" from the popup menu.

The operator will automatically compile and build the required slmeta
file for the shader and then attempt to build a shader panel in the active
pipeline.

this fixes #22 issue on GitHub repo: https://github.com/nfz/RIBMosaic-exp/issues/22