ExportWorld class implemented:
ExportWorld now has an export_rib() method that takes care of exporting
all the objects in the world.  ExportPass delegates to ExportWorld
the handling of object exporting.  The world can be setup as a ReadArchive.

World object can now have image and volume shaders attached to it and will
export them.