Bug Fix: Preview was not updating after pipeline sync executed:
If the user loaded a blend file that was using RIBMosaic and selected
"sync pipelines" the preview would not update unless forced by the user.

The WM_OT_ribmosaic_modal_sync() operator renamed to WM_OT_ribmosiac_pipeline_sync()
to better reflect what it does.  It has also been promoted to a full operator
for the pipeline manager.