Blender Material Preview now updates when a shader parameter is modified by the user.
In Blender 2.57 the user had to force a material preview update by selecting one
of the preview options.

Blender 2.58 now supports property callbacks which allows a python script to perform
some action when the property value changes.  Shader parameters that are set
up as a property on the shader panel now have a callback assigned which does
what the user did in 2.57: force a preview update by changing the preview render type.