New Feature: Ribmosaic panels can now be associated with a pass or group of passes:
- Each panel now has a property called "Pass Filter" that allows the user to
associate the panel with a specific pass, all passes, or a group of passes.
Its just a string that contains a list of the pass names that the panel is to
be used.  If the string is empty then the panel works with all passes.  Pass
names are seperated by a comma if more than one name is in the string.

- Bug Fix: Lights will now not be exported to RIB when RIB archive property
 in Export Options panel is set to No Export.
This fixes #31 issue on GitHub Repo: https://github.com/nfz/RIBMosaic-exp/issues/31