Shadow maps can now be generated using the ZFile display in Aqsis:
The zfile display utility panel can now be attached to render settings and be
made to only be active for a specific pass ie shadow map generation pass.
The utility will automatically add the MakeShadow rib statement at the end of
the rib in order to converter the image to a proper shadow map.